calculator/requirements.txt
grpcio-tools ~= 1.30
numpy